requirements.txt
streamlit
psycopg2-binary
pandas
bcrypt
numpy
opencv-python-headless
pyzbar
